Fix for CB that allows more than 1 DOP when imported DB

+++ test/unit/test_devops.py
@@ -83,7 +83,8 @@ def test_count() -> None:
     assert devops.count(tutil.SQ, "bitbucket") == 1
     assert devops.count(tutil.SQ, "bitbucketcloud") == 1
     # TODO: Find out if normal than multiple devops platforms allowed on CE
-    nb_gh = 2 if tutil.SQ.edition() in (c.EE, c.DCE) else 1
+    # nb_gh = 2 if tutil.SQ.edition() in (c.EE, c.DCE) else 1
+    nb_gh = 2
     assert devops.count(tutil.SQ, "github") == nb_gh
     assert devops.count(tutil.SQ) == nb_gh + 4
 
@@ -123,11 +124,24 @@ def test_import_config_2() -> None:
         with pytest.raises(exceptions.UnsupportedOperation):
             devops.import_config(endpoint=tutil.SQ, config_data=dop)
         return
-    assert devops.import_config(endpoint=tutil.SQ, config_data=dop) == 3
-    assert devops.exists(endpoint=tutil.SQ, key="ADO2")
-    assert devops.exists(endpoint=tutil.SQ, key="GH2")
-    assert devops.exists(endpoint=tutil.SQ, key="GL2")
-    for key in "ADO2", "GH2", "GL2":
-        obj = devops.get_object(endpoint=tutil.SQ, key=key)
+    if tutil.SQ.edition() in (c.CE, c.DE):
+        counts = {"azure": devops.count(tutil.SQ, "azure"), "gitlab": devops.count(tutil.SQ, "gitlab"), "github": devops.count(tutil.SQ, "github")}
+    else:
+        counts = {"azure": 0, "gitlab": 0, "github": 0}
+    deduction = min(1, counts["azure"]) + min(1, counts["gitlab"]) + min(1, counts["github"])
+    assert devops.import_config(endpoint=tutil.SQ, config_data=dop) == 3 - deduction
+    if counts["azure"] == 0:
+        assert devops.exists(endpoint=tutil.SQ, key="ADO2")
+        obj = devops.get_object(endpoint=tutil.SQ, key="ADO2")
         obj.delete()
-        assert devops.exists(endpoint=tutil.SQ, key=key) is False
+        assert not devops.exists(endpoint=tutil.SQ, key="ADO2")
+    if counts["github"] == 0:
+        assert devops.exists(endpoint=tutil.SQ, key="GH2")
+        obj = devops.get_object(endpoint=tutil.SQ, key="GH2")
+        obj.delete()
+        assert not devops.exists(endpoint=tutil.SQ, key="GH2")
+    if counts["azure"] == 0:
+        assert devops.exists(endpoint=tutil.SQ, key="GL2")
+        obj = devops.get_object(endpoint=tutil.SQ, key="GL2")
+        obj.delete()
+        assert not devops.exists(endpoint=tutil.SQ, key="GL2")
